requirements.txt
opencv-python
paho-mqtt
pymongo
